test: 이벤트 운영 기간 테스트 코드 작성
